net: dwc_eth_qos: Use DMA allocator for descriptors
Per the comment above eqos_alloc_descs(), the DMA descriptors should
be allocated from coherent/non-cached memory if available. On some
platforms, coherent memory must be accessed using a different virtual
address on the CPU than the handle passed to the device. Use the DMA
allocator for the descriptors, which abstracts away this detail from the
driver. If coherent memory is unavailable, the DMA allocator is just a
wrapper around memalign()/free(), so there is no functional change for
most platforms.

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>