dma-mapping: Provide a generic dma_{alloc,free}_coherent()
This is a combination of existing architecture-specific implementations.
Using malloc_cache_aligned() includes the fix from commit 9d47d1316da6
("arm: Round the dma_alloc_coherent memory size to cache line aligned")
that is currently missing on riscv and x86. For architectures that do
not define MAP_NOCACHE, this is equivalent to the current arm/riscv/x86
implementation, with the addition of the dcache invalidation. However,
the same implementation allows architectures (or specific platforms) to
define MAP_NOCACHE and get non-cacheable memory.

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>